libbomp: merged xomp and bomp together and refactoration
There may be a use case where the bomp and xomp library are used
together i.e. we have a linker conflict and not the proper functions are
called.

Refactoration: instead of global variables there is a single state
pointer which can be exchanged depending on the backend to be used.
Further function pointers in this state ensure that the correct
functions are called.

added more sub functions for GCC pragma function calls.

moved omp.h into the global include directory.

merged: bomp_init with bomp_span_domain

+++ lib/bomp/include/spin.h
@@ -20,8 +20,7 @@
 #include <omp.h>
 
 /** \brief spinlock */
-// why 64-bit? this means we need an extra prefix on the lock ops... -AB
-typedef volatile unsigned long bomp_lock_t;
+typedef volatile unsigned int bomp_lock_t;
 
 static inline void bomp_lock(bomp_lock_t *lock)
 {
